+++ Util/grasp_to_gspbox.m
@@ -6,11 +6,15 @@
 %
 % Authors:
 %  - Benjamin Girault <benjamin.girault@ens-lyon.fr>
+%  - Benjamin Girault <benjamin.girault@usc.edu>
 
 % Copyright Benjamin Girault, École Normale Supérieure de Lyon, FRANCE /
-% Inria, FRANCE (2015-11-01)
+% Inria, FRANCE (2015-2016)
+% Copyright Benjamin Girault, University of Sourthern California, Los
+% Angeles, California, USA (2018)
 % 
 % benjamin.girault@ens-lyon.fr
+% benjamin.girault@usc.edu
 % 
 % This software is a computer program whose purpose is to provide a Matlab
 % / Octave toolbox for handling and displaying graph signals.
@@ -64,5 +68,6 @@ function gsp_graph = grasp_to_gspbox(graph)
     gsp_graph.grasp_background = graph.background;
     gsp_graph.type = 'GraSP';
     
+    grasp_start_opt_3rd_party('gspbox');
     gsp_graph = gsp_graph_default_parameters(gsp_graph);
 end